+++ doc/CHANGELOG.md
@@ -32,6 +32,7 @@ The full set of tests is only run for tagged releases.
     - librtools/Rstats: add Reset(); drop operator-=() and operator*=()
     - librtcltools/RtclStats: Rename Collect->Exec, not longer const; add -reset
     - libr*/*: Stats() not longer const; use RtclStats::Exec()
+    - librtools/RparseUrl: add DirName,FileName,FileStem,FileType
 - firmware changes
   - sys_w11a_arty: down-rate to 72 MHz, viv 2019.1 fails with 75 MHz
   - sys_w11a_*.vmfset: add new rule for vivado 2019.1